Consolidated css/js into corresponding dirs. Dropping modules.
