+++ gruntconfig/browserify.js
@@ -13,16 +13,8 @@ var browserify = {
       paths: [
         CWD + '/' + config.src + '/htdocs/js',
         CWD + '/' + config.src + '/htdocs/modules',
-        NODE_MODULES + '/earthquake-eventpages/src/htdocs/js',
-        NODE_MODULES + '/earthquake-eventpages/src/htdocs/modules',
-        NODE_MODULES + '/hazdev-accordion/src',
         NODE_MODULES + '/hazdev-cache-invalidator/src/htdocs/js',
-        NODE_MODULES + '/hazdev-location-view/src',
-        NODE_MODULES + '/hazdev-question-view/src',
-        NODE_MODULES + '/hazdev-svgimagemap/src',
-        NODE_MODULES + '/hazdev-tablist/src',
-        NODE_MODULES + '/hazdev-webutils/src',
-        NODE_MODULES + '/quakeml-parser-js/src'
+        NODE_MODULES + '/hazdev-webutils/src'
       ]
     }
   },
