+++ src/htdocs/css/_EditProductView.scss
