+++ src/htdocs/css/products.scss
@@ -1,10 +1,10 @@
 @import 'hazdev-webutils/src/mvc/ModalView.scss';
 
-@import 'modules/admin/_AllProductsPage.scss';
-@import 'modules/admin/_EditLinkView.scss';
-@import 'modules/admin/_EditProductView.scss';
-@import 'modules/admin/_ProductActionsView.scss';
-@import 'modules/admin/_SendProductView.scss';
+@import '_AllProductsPage.scss';
+@import '_EditLinkView.scss';
+@import '_EditProductView.scss';
+@import '_ProductActionsView.scss';
+@import '_SendProductView.scss';
 
 @import '_FileUploadStatusView.scss';
 @import '_ContentsManagerView.scss';
