+++ src/htdocs/js/admin/AssociateEventView.js
