+++ src/htdocs/js/admin/EditLinkView.js
