+++ src/htdocs/js/admin/EditProductView.js
