+++ src/htdocs/js/admin/EventComparisonView.js
