+++ src/htdocs/js/admin/EventsNearbyView.js
