+++ src/htdocs/js/admin/ProductDetailsView.js
