+++ src/htdocs/js/admin/ProductFactory.js
