+++ src/htdocs/js/admin/SendProductView.js
