+++ src/htdocs/js/admin/TextProductView.js
