--- src/htdocs/modules/admin/_AdminSummaryDetailsPage.scss
@@ -1,10 +1,0 @@
-.edit-summary {
-
-  background-color: #fff;
-  margin: 1em 0;
-
-  > .product-action-view {
-    margin: 0;
-  }
-
-}
