--- src/htdocs/modules/admin/_AdminSummaryPage.scss
@@ -1,3 +1,0 @@
-.invalidator-view #paths {
-  height: 10em;
-}
