--- src/htdocs/modules/admin/_AssociateEventView.scss
@@ -1,5 +1,0 @@
-
-.modal-content .textproduct-text {
-  width: 100%;
-  height: 10em;
-}
