--- src/htdocs/modules/admin/_EventComparisonView.scss
@@ -1,22 +1,0 @@
-table.event-comparison > tbody > tr > td {
-  vertical-align:top;
-}
-
-td.actions > button {
-  margin: 4px auto;
-
-  & + button {
-    margin-left: 1em;
-  }
-}
-
-small.value-diff {
-  color: #999;
-  display:block;
-  font-size: 12px;
-  font-weight: 300;
-}
-
-tr.reference-event > td {
-  background-color: rgba(231, 240, 250, 0.5);
-}
