--- src/htdocs/modules/admin/_LinkSummaryDetailsView.scss
@@ -1,3 +1,0 @@
-.link-summary > dl {
-  margin-top: 0;
-}
