--- src/htdocs/modules/admin/_ProductHistoryView.scss
@@ -1,29 +1,0 @@
-
-.product-history {
-
-  > .alert {
-    background-color: #fff;
-    margin: 1em 0;
-
-    > .metadata {
-      color: #666;
-      float: right;
-      font-size: .9em;
-    }
-
-    > .product-action-view {
-      margin: 0;
-    }
-  }
-}
-
-
-.deleted > td,
-.deleted > ul > li > span {
-  text-decoration: line-through;
-}
-
-.superseded > ul > li > span {
-  color: #999;
-  font-weight: 300;
-}
