--- src/htdocs/modules/admin/_TextSummaryDetailsView.scss
@@ -1,4 +1,0 @@
-.text-summary {
-  border: 1px solid #ddd;
-  padding: 8px;
-}
