--- src/htdocs/modules/admin/index.scss
@@ -1,19 +1,0 @@
-
-@import '_AdminSummaryDetailsPage.scss';
-@import '_AdminSummaryPage.scss';
-@import '_AssociateEventView.scss';
-@import '_AllProductsPage.scss';
-@import '_EditLinkView.scss';
-@import '_EditProductView.scss';
-@import '_EventComparisonView.scss';
-@import '_LinkSummaryDetailsView.scss';
-@import '_ProductActionsView.scss';
-@import '_ProductHistoryView.scss';
-@import '_SendProductView.scss';
-@import '_TextSummaryDetailsView.scss';
-
-label > small {
-  display:block;
-  color: #666;
-  font-size: smaller;
-}
