--- src/htdocs/modules/admin/scientific/_ScientificSummaryPage.scss
@@ -1,11 +1,0 @@
-.edit-link {
-  border: 1px solid #ddd;
-  padding: 1em;
-  margin: 1em 0 0;
-
-  > .product-action-view {
-    display: block;
-    margin: 0;
-  }
-
-}
