--- src/htdocs/modules/admin/scientific/index.scss
@@ -1,1 +1,0 @@
-@import '_ScientificSummaryPage.scss';
