--- src/htdocs/modules/admin/summary/_SummaryPage.scss
@@ -1,29 +1,0 @@
-.edit-link {
-  &.deleted > a {
-    text-decoration: line-through;
-  }
-}
-
-.edit-text {
-  margin-top: 1em;
-  padding: 1em;
-  border: 1px solid #ddd;
-
-  > .product-action-view {
-    margin: 0;
-  }
-
-  &.deleted > section {
-    text-decoration: line-through;
-  }
-}
-
-.add-product-button {
-  border-left: 5px solid #3c3;
-
-  @media screen and (max-width: 767px) {
-    font-size: .8125em;
-    line-height: 3rem;
-    height: 3rem;
-  }
-}
