--- src/htdocs/modules/admin/summary/index.scss
@@ -1,6 +1,0 @@
-@import '_SummaryPage.scss';
-
-
-.deleted .origin-detail {
-  text-decoration: line-through;
-}
